entities2

+++ routes/tokenvortex/api/entities/delete.js
@@ -4,13 +4,18 @@ var cors = require('cors');
 router.use(cors());
 var bodyParser = require("body-parser");
 var EntitiesModel = require('../../models/entities');
+const admin_id = '5f8f88e5d28b37394459bbba';
 
 
 
 
 //delete by id
 router.post("/:_id", bodyParser.json(), async function(req, res, next) {
-  let result =  await ContractsModel.findByIdAndDelete(_id).exec();
+  let _id = req.params._id
+  let result = {};
+  if(_id !== admin_id){
+    result =  await EntitiesModel.findByIdAndDelete(_id).exec();
+  }
   return res.send(result);
   });
   
@@ -18,9 +23,11 @@ router.post("/:_id", bodyParser.json(), async function(req, res, next) {
 //soft delete
   router.post("/soft/:_id", bodyParser.json(),async  function(req, res, next) {
     let _id = req.params._id;
-    ContractsModel.updateOne({ _id }, { isActive: false }).then(result => {
+    if(_id !== admin_id){
+      EntitiesModel.updateOne({ _id }, { isActive: false }).then(result => {
       return res.send(result);
     });
+  }
   });
   
 
@@ -29,9 +36,12 @@ router.post("/:_id", bodyParser.json(), async function(req, res, next) {
 router.post("/return/:_id", bodyParser.json(),async  function(req, res, next) {
   let _id = req.params._id;
   let user_id = req.user_id;
+  if(_id !== admin_id){
+
   await EntitiesModel.deleteOne({ _id}).exec();
   let entities = EntitiesModel.find({user_id}).exec();
   res.send(entities);
+  }
 });
 
 module.exports = router;
